Run validator for DeleteProductCommand in the handler

+++ Application/Features/Products/Commands/Delete/DeleteProductCommandHandler.cs
@@ -1,4 +1,5 @@
 ﻿using Application.Services.Products;
+using FluentValidation;
 using MediatR;
 
 namespace Application.Features.Products.Commands.Delete
@@ -8,13 +9,21 @@ namespace Application.Features.Products.Commands.Delete
 
         private readonly IProductService _productService;
 
-        public DeleteProductCommandHandler(IProductService productService)
+        private readonly IValidator<DeleteProductCommand> _validator;
+
+        public DeleteProductCommandHandler(IProductService productService, IValidator<DeleteProductCommand> validator)
         {
             _productService = productService;
+            _validator = validator;
         }
 
         public async Task Handle(DeleteProductCommand request, CancellationToken cancellationToken)
         {
+            var validationResult = await _validator.ValidateAsync(request);
+
+            if (!validationResult.IsValid)
+                throw new ValidationException(validationResult.Errors);
+
             await _productService.DeleteProduct(request.UserId, request.ProductId);
         }
